Cambios para acoplar is_taxable

+++ Steady-Management.DataAccess/ProductData.cs
@@ -58,7 +58,7 @@ namespace Steady_Management.Data
                     rdr.GetInt32(rdr.GetOrdinal("product_id")),
                     rdr.GetInt32(rdr.GetOrdinal("category_id")),
                     rdr.GetString(rdr.GetOrdinal("product_name")),
-                    (decimal)rdr.GetDouble(rdr.GetOrdinal("price")),
+                    (decimal)rdr.GetDecimal(rdr.GetOrdinal("price")),
                     rdr.GetBoolean(rdr.GetOrdinal("is_taxable"))
                 ));
             }
